# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GabFlow API
  version: 0.4.0
  description: API principal para gestão de gabinetes municipais.
servers:
  - url: https://api.gabflow.example/api/v1
security:
  - cookieAuth: []
paths:
...
  /solicitacoes/{id}/tarefas:
    post:
      summary: Cria uma tarefa vinculada à solicitação
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [titulo]
              properties:
                titulo: { type: string, minLength: 3, maxLength: 180 }
                descricao: { type: string }
                responsavelId: { type: string, format: uuid }
                prioridade: { type: string, enum: [BAIXA, MEDIA, ALTA, CRITICA] }
                prazo: { type: string, format: date-time }
      responses:
        "201":
          description: Tarefa criada
  /tarefas/{id}:
    patch:
      summary: Atualiza estado, prioridade ou prazo da tarefa
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tarefa atualizada
  /solicitacoes/agrupar-duplicadas:
    post:
      summary: Agrupa solicitações duplicadas sem excluí-las
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [solicitacaoIds, motivo]
              properties:
                solicitacaoIds:
                  type: array
                  minItems: 2
                  items: { type: string, format: uuid }
                motivo: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "201":
          description: Grupo de duplicidades criado ou ampliado
  /solicitacoes/{id}/anexos:
    post:
      summary: Envia e verifica um anexo da solicitação
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [arquivo]
              properties:
                arquivo: { type: string, format: binary }
      responses:
        "201":
          description: Anexo validado e armazenado
...
  /anexos/{id}/download:
    get:
      summary: Baixa um anexo por token temporário assinado
      parameters:
        - $ref: "#/components/parameters/Id"
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Conteúdo do anexo
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...